requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.12.1
bcrypt==4.0.1
certifi==2026.2.25
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
passlib==1.7.4
pgvector==0.4.2
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
